Began Implementing Problem Functions
Started to implement methods for working with the Problem struct,
including adding variables and constraints, as well as creating new
models. Added a dependency for cfg_if which will be used in the future
to handle different solver features being enabled.

TODO: Add functions for removing variables and constraints from problems